app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
asyncpg==0.27.0
databases==0.6.0
SQLAlchemy==1.4.46
pydantic==1.10.5
psycopg2-binary==2.9.6
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
twilio==9.2.0
pydantic[email]==1.10.5